+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `wrapperClassname` prop to `EXPERIMENTAL_Table/Sections`.
+
+### Fixed
+
+- `EXPERIMENTAL_Table` height.
+
 ## [9.127.0] - 2020-08-13
 
 ### Added